camera bug
