+++ lib/main.dart
@@ -12,30 +12,31 @@ import 'camera.dart' as camera;
 import 'facts.dart' as facts;
 import 'image_process.dart' as image_process;
 import 'description_process.dart' as description_process;
+import 'home.dart' as home;
 
-//void main() => runApp(MyApp());
+void main() => runApp(MyApp());
 
-Future<void> main() async {
-  // Ensure that plugin services are initialized so that `availableCameras()`
-  // can be called before `runApp()`
-  WidgetsFlutterBinding.ensureInitialized();
 
-  // Obtain a list of the available cameras on the device.
-  final cameras = await availableCameras();
-
-  // Get a specific camera from the list of available cameras.
-  final firstCamera = cameras.first;
-
-  runApp(
-    MaterialApp(
+class MyApp extends StatelessWidget {
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
       title: 'Bear Grylls',
       theme: ThemeData(
+        // This is the theme of your application.
+        //
+        // Try running your application with "flutter run". You'll see the
+        // application has a blue toolbar. Then, without quitting the app, try
+        // changing the primarySwatch below to Colors.green and then invoke
+        // "hot reload" (press "r" in the console where you ran "flutter run",
+        // or simply save your changes to "hot reload" in a Flutter IDE).
+        // Notice that the counter didn't reset back to zero; the application
+        // is not restarted.
         primarySwatch: Colors.blue,
       ),
-      home: camera.TakePictureScreen(
-          camera: firstCamera
-      ),
-
-    )
-  );
+      home: home.HomePage(),
+    );
+  }
 }
+
